+++ src/d_stream/mod.rs
@@ -19,11 +19,11 @@
 // er General Public License along with bzipper. If
 // not, see <https://www.gnu.org/licenses/>.
 
-use crate::error::{Error, Result};
+use crate::{Error, Result, SStream};
 
 use std::fmt::{Debug, Formatter};
 
-/// A byte stream for deserialisation.
+/// Byte stream for deserialisation.
 ///
 /// This type borrows a byte slice (hence [`new`](DStream::new)), keeping track internally of the used bytes.
 #[derive(Clone)]
@@ -34,6 +34,8 @@ pub struct DStream<'a> {
 
 impl<'a> DStream<'a> {
 	/// Constructs a new byte stream.
+	#[inline(always)]
+	#[must_use]
 	pub fn new<T: AsRef<[u8]> + ?Sized>(buf: &'a T) -> Self { Self {
 		data: buf.as_ref(),
 		len:  buf.as_ref().len(),
@@ -54,16 +56,31 @@ impl<'a> DStream<'a> {
 
 		Ok(&self.data[start..stop])
 	}
+
+	/// Takes a slice of the remaining data.
+	#[must_use]
+	pub fn as_slice(&self) -> &[u8] {
+		let stop  = self.data.len();
+		let start = stop - self.len;
+
+		&self.data[start..stop]
+	}
+
+	/// Converts the stream to a `SStream` object.
+	///
+	/// The returned object owns a copy of the remaining data.
+	#[inline(always)]
+	#[must_use]
+	pub fn to_s_stream(&self) -> SStream {
+		SStream(self.as_slice().to_vec())
+	}
 }
 
 impl Debug for DStream<'_> {
 	fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
-		let stop  = self.data.len();
-		let start = self.data.len() - self.len;
-
 		write!(f, "[")?;
 
-		for v in &self.data[start..stop] { write!(f, "{v:#02X},")? };
+		for v in self.as_slice() { write!(f, "{v:#02X},")? };
 
 		write!(f, "]")?;