+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.1.0
+
+* Bump minor
+* Export all in crate root
+* Add fixed string type
+* Add new errors
+* Update documentation
+* Add `as_d_stream` method to `SStream`
+* Add `to_s_stream` and `as_slice` methods to `DStream`
+
 # 0.0.2
 
 * Add license files